LPS-73140 Moves constant to the constants package for consistency
