+++ modules/apps/foundation/portal-template/portal-template-soy/src/main/java/com/liferay/portal/template/soy/utils/SoyContext.java
@@ -14,6 +14,8 @@
 
 package com.liferay.portal.template.soy.utils;
 
+import com.liferay.portal.template.soy.constants.SoyTemplateConstants;
+
 import java.util.HashMap;
 import java.util.Map;
